requirements.txt
playwright==1.40.0
pytest==7.4.0
pytest-playwright==0.4.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-base-url==2.1.0